Remove explicit coordinates and move class to interfaces
